Change RecognizerMatch to use multiple ranges
